src-testbed/requirements.txt
Flask==1.1.2
Flask-BasicAuth==0.2.0
Flask-RESTful==0.3.8
numpy==1.18.4
requests==2.22.0
Pillow
grpcio==1.29.0
grpcio-tools==1.29.0
gunicorn==20.0.4
redis==3.5.3
gevent==20.9.0
greenlet==0.4.17
locust==1.0.3
tornado
docker
pycurl
wrapt
pandas